+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This log documents the significant changes for each release.
 This project follows [Semantic Versioning](http://semver.org/).
 
 ## [6.0.4] - 2022-12-21
+### Changed
+- Replaced calls to LiveAnnouncer with calls to Def.ScreenReaderLog from
+  autocomplete-lhc.
 ### Fixed
 - Issues with initialization spinner:
   - there were no announcements for the screen reader,